Refactor stack values
