+++ .vm/src/program/stack/circuit_value/to_bits.rs
@@ -14,19 +14,26 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-mod to_bits;
-mod to_fields;
+use super::*;
 
-use console::{
-    network::prelude::*,
-    program::{Plaintext, Record},
-    types::Field,
-};
+impl<A: circuit::Aleo> ToBits for CircuitValue<A> {
+    type Boolean = circuit::types::Boolean<A>;
 
-#[derive(Clone, PartialEq, Eq)]
-pub enum StackValue<N: Network> {
-    /// A plaintext value.
-    Plaintext(Plaintext<N>),
-    /// A record value.
-    Record(Record<N, Plaintext<N>>),
+    /// Returns the circuit value as a list of **little-endian** bits.
+    #[inline]
+    fn to_bits_le(&self) -> Vec<circuit::types::Boolean<A>> {
+        match self {
+            CircuitValue::Plaintext(plaintext) => plaintext.to_bits_le(),
+            CircuitValue::Record(record) => record.to_bits_le(),
+        }
+    }
+
+    /// Returns the circuit value as a list of **big-endian** bits.
+    #[inline]
+    fn to_bits_be(&self) -> Vec<circuit::types::Boolean<A>> {
+        match self {
+            CircuitValue::Plaintext(plaintext) => plaintext.to_bits_be(),
+            CircuitValue::Record(record) => record.to_bits_be(),
+        }
+    }
 }
