+++ .vm/src/program/stack/stack_value/to_bits.rs
@@ -14,19 +14,24 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-mod to_bits;
-mod to_fields;
+use super::*;
 
-use console::{
-    network::prelude::*,
-    program::{Plaintext, Record},
-    types::Field,
-};
+impl<N: Network> ToBits for StackValue<N> {
+    /// Returns the stack value as a list of **little-endian** bits.
+    #[inline]
+    fn to_bits_le(&self) -> Vec<bool> {
+        match self {
+            StackValue::Plaintext(plaintext) => plaintext.to_bits_le(),
+            StackValue::Record(record) => record.to_bits_le(),
+        }
+    }
 
-#[derive(Clone, PartialEq, Eq)]
-pub enum StackValue<N: Network> {
-    /// A plaintext value.
-    Plaintext(Plaintext<N>),
-    /// A record value.
-    Record(Record<N, Plaintext<N>>),
+    /// Returns the stack value as a list of **big-endian** bits.
+    #[inline]
+    fn to_bits_be(&self) -> Vec<bool> {
+        match self {
+            StackValue::Plaintext(plaintext) => plaintext.to_bits_be(),
+            StackValue::Record(record) => record.to_bits_be(),
+        }
+    }
 }
